+++ packages/opentelemetry-context-async-hooks/src/AbstractAsyncHooksContextManager.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { ContextManager, Context } from '@opentelemetry/api';
+import type { ContextManager, Context, Token } from '@opentelemetry/api';
 import { EventEmitter } from 'events';
 
 type Func<T> = (...args: unknown[]) => T;
@@ -40,6 +40,10 @@ export abstract class AbstractAsyncHooksContextManager
   abstract enable(): this;
 
   abstract disable(): this;
+
+  abstract attach(context: Context): Token;
+
+  abstract detach(token: Token): void;
 
   /**
    * Binds a the certain context or the active one to the target function and then returns the target